Fixed print.survHE issue for GenGamma ~ 1
Fixed print.survHE issue for Generalized Gamma model fitted with no covariates, i.e. Surv(...) ~ 1.
When the table object was subset with only one beta, the second dimension of the matrix was dropped but still subset on subsequently. Added a drop=FALSE in the subset by grep("beta") to avoid it.